+++ sources/player.hpp
@@ -24,7 +24,7 @@ class Player{
     //Functions:
     void setScoreSize(int n);
     int stacksize();//prints the amount of cards left. should be 21 but can be less if a draw was played
-    int cardesTaken();// prints the amount of cards this player has won.
+    double cardesTaken();// prints the amount of cards this player has won.
     
 };
 };
